pyproject.toml
[project]
name = "data-sharding"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faker>=37.8.0",
    "fastapi[standard]>=0.100.0",
    "numpy>=2.3.3",
    "psycopg2>=2.9.10",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "pytest>=8.4.2",
]
